backend/requirements.txt
asgiref==3.11.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
cryptography==46.0.5
dj-database-url==3.1.2
dj-rest-auth==7.2.0
Django==6.0.3
django-allauth==65.15.0
django-cors-headers==4.9.0
djangorestframework==3.17.0
djangorestframework_simplejwt==5.5.1
idna==3.11
psycopg2-binary==2.9.11
pycparser==3.0
PyJWT==2.12.1
python-decouple==3.8
requests==2.32.5
sqlparse==0.5.5
urllib3==2.6.3
